requirements.txt
APScheduler==3.8.1
attrs==21.2.0
azure-common==1.1.27
azure-core==1.20.1
azure-nspkg==3.0.2
azure-storage-blob==12.9.0b1
backports.zoneinfo==0.2.1
boto3==1.20.4
botocore==1.23.4
Bottleneck==1.3.2
cachetools==4.2.4
cassandra-driver==3.25.0
category-encoders==2.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
dpcpp-cpp-rt==2021.4.0
et-xmlfile==1.1.0
Flask==2.0.2
from-root==1.0.2
geomet==0.2.1.post1
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.8.1
imblearn==0.0
intel-cmplr-lib-rt==2021.4.0
intel-cmplr-lic-rt==2021.4.0
intel-opencl-rt==2021.4.0
intel-openmp==2021.4.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib==3.4.3
missingno==0.5.0
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msrest==0.6.21
multimethod==1.4
mysql-connector-python==8.0.27
networkx==2.6.3
numexpr==2.7.3
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.4
pandas-profiling==3.0.0
patsy==0.5.2
phik==0.12.0
Pillow==8.4.0
plotly==5.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
schedule==1.1.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.27
statsmodels==0.13.0
tangled-up-in-unicode==0.1.0
tbb==2021.4.0
tenacity==8.0.1
threadpoolctl==2.2.0
tqdm==4.62.3
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
visions==0.7.1
Werkzeug==2.0.2
win32-setctime==1.0.3
wincertstore==0.2
supervisor==4.2.2